refactor(scripts): give the shared constants and helpers one home
`CHARTS_DIR = Path("charts")` was declared in seven files of the contract group,
`FIRST_PARTY` in two, `_read_yaml` — one line, and the `or {}` on it is the whole
reason it is a function — in three, and `dig` byte-identical in two, with only one
of the pair carrying the docstring.

None of that was hurting anything, which is the argument for fixing it now rather
than after it does: an identical declaration in seven files is a fact that can be
changed in one of them and stay wrong in the other six, silently, because every
copy still parses and every test still passes. `FIRST_PARTY` is the one where that
would bite hardest — `check-contract-coverage` and `new-chart` have to agree about
which images are first-party, or the scaffold writes charts the gate rejects.

`config_paths` deliberately holds nothing that decides anything. `config_contract`
owns the contract model, `config_declaration` owns the declaration, and a helper
that grows a rule belongs in whichever of those the rule is about. What lives here
is "where is charts/" and "read this YAML", which are not rules and had no owner.

`config_declaration.dig` keeps its name by importing it, since
`generate-contract-tests` imports it from there.

`check-immutable-fields.py` still declares its own `CHARTS_DIR` and is left alone:
it imports nothing from this directory and has no `sys.path` bootstrap, so sharing
a one-line constant would mean adding one to a script outside this group.
