+++ .github/scripts/contract-diff.py
@@ -53,9 +53,9 @@ import yaml
 sys.path.insert(0, str(Path(__file__).resolve().parent))
 
 import config_diff as cd  # noqa: E402
+from config_paths import CHARTS_DIR  # noqa: E402
 from config_report import Report, error  # noqa: E402
 
-CHARTS_DIR = Path("charts")
 DEFAULT_REF = "origin/main"
 
 # The JSON's own version, on the same reasoning as `terrace_contract`: a consumer that does not
